+++ cases/testcases/dynamic_earmuffs/no_dynamic/red.clj
@@ -1,0 +1,3 @@
+(ns testcases.dynamic-earmuffs.no-dynamic.red)
+
+(def *foo*)