+++ cases/testcases/dynamic_earmuffs/no_earmuffs/green.clj
@@ -1,0 +1,3 @@
+(ns testcases.dynamic-earmuffs.no-earmuffs.green)
+
+(def foo)